Merge

+++ src/main/java/TRaMis8khae/starbucks/product/domain/SubCategory.java
@@ -1,12 +1,13 @@
 package TRaMis8khae.starbucks.product.domain;
 
 import jakarta.persistence.*;
+import jakarta.persistence.criteria.CriteriaBuilder;
 
 @Entity
 public class SubCategory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long subCategoryId;
+    private Integer subCategoryId;
     @Column(nullable = false, length = 50)
     private String subCategoryName;
     private Integer subCategoryOrder;
